Ray casting ok

+++ pikuma/cub3d/includes/player.h
@@ -6,7 +6,7 @@
 /*   By: sbeylot <sbeylot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/31 07:32:50 by sbeylot           #+#    #+#             */
-/*   Updated: 2023/01/02 09:35:21 by sbeylot          ###   ########.fr       */
+/*   Updated: 2023/01/03 15:05:31 by sbeylot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,6 @@ t_player	*init_player(t_cub3d *cub);
 void		draw_player(t_cub3d *cub);
 void	update_player(t_cub3d *cub);
 
-bool	is_a_wall(t_cub3d *cub, int x, int y);
+bool	is_a_wall(t_cub3d *cub, double x, double y);
 
 #endif
